refactor(ui): :recycle: extract to separate components
