+++ src/components/index.ts
@@ -1,5 +1,3 @@
-export * from './ClassDetail'
-export * from './ClassStatusSwitch'
 export * from './ClassTable'
 export * from './HeaderBar'
 export * from './SideNav'
